+++ src/Ryujinx.Ava/UI/ViewModels/Input/ControllerInputViewModel.cs
@@ -2,16 +2,11 @@ using Avalonia.Svg.Skia;
 using Ryujinx.Ava.UI.Models;
 using Ryujinx.Ava.UI.Views.Input;
 
-namespace Ryujinx.Ava.UI.ViewModels
+namespace Ryujinx.Ava.UI.ViewModels.Input
 {
     public class ControllerInputViewModel : BaseModel
     {
         private ControllerInputConfig _config;
-        private bool _isLeft;
-        private bool _isRight;
-        private bool _showSettings;
-        private SvgImage _image;
-
         public ControllerInputConfig Config
         {
             get => _config;
@@ -22,6 +17,7 @@ namespace Ryujinx.Ava.UI.ViewModels
             }
         }
 
+        private bool _isLeft;
         public bool IsLeft
         {
             get => _isLeft;
@@ -32,6 +28,7 @@ namespace Ryujinx.Ava.UI.ViewModels
             }
         }
 
+        private bool _isRight;
         public bool IsRight
         {
             get => _isRight;
@@ -42,6 +39,7 @@ namespace Ryujinx.Ava.UI.ViewModels
             }
         }
 
+        private bool _showSettings;
         public bool ShowSettings
         {
             get => _showSettings;
@@ -52,6 +50,7 @@ namespace Ryujinx.Ava.UI.ViewModels
             }
         }
 
+        private SvgImage _image;
         public SvgImage Image
         {
             get => _image;
