+++ src/Ryujinx.Ava/UI/ViewModels/Input/MotionInputViewModel.cs
